+++ src/Restaurant.ts
@@ -1,13 +1,4 @@
-export type TCategory = '한식' | '중식' | '일식' | '아시안' | '양식' | '기타';
-type TDistance = 5 | 10 | 15 | 20 | 30;
-
-export interface IRestaurant {
-  category: TCategory;
-  name: string;
-  distance: TDistance;
-  description?: string;
-  referenceLink?: string;
-}
+import type { IRestaurant } from './types/restaurant';
 
 class Restaurant {
   information;
